+++ trunk/lib/client/CServerProxy.cpp
@@ -771,7 +771,7 @@ CServerProxy::setOptions()
 	m_client->setOptions(options);
 
 	// update modifier table
-	for (UInt32 i = 0, n = options.size(); i < n; i += 2) {
+	for (UInt32 i = 0, n = (UInt32)options.size(); i < n; i += 2) {
 		KeyModifierID id = kKeyModifierIDNull;
 		if (options[i] == kOptionModifierMapForShift) {
 			id = kKeyModifierIDShift;
